requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
supabase>=2.3.5,<=2.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic[email]>=2.9.0
httpx>=0.27.0
